+++ src/data/kaysaar/aotd/tot/plugins/AoTDToolboxTheoryPlugin.java
@@ -400,4 +400,3 @@ public class AoTDToolboxTheoryPlugin extends BaseModPlugin implements MarketCont
 
 
 
-